requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.3
django-allauth==0.54.0
django-cors-headers==4.2.0
django-filter==23.2
django-translations==1.3.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
idna==3.4
oauthlib==3.2.2
Pillow==10.0.0
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.0.4
